Move hostKernel to utilio.HostKernel and deduplicate
hostKernel was duplicated in both cmd and nodeupdate packages with
identical implementations. Consolidate into a single exported
utilio.HostKernel() function and update both callers.
