+++ cmd/agent/internal/utilio/hostkernel_linux.go
@@ -3,7 +3,7 @@
 
 //go:build linux
 
-package nodeupdate
+package utilio
 
 import (
 	"fmt"
@@ -11,13 +11,15 @@ import (
 	"syscall"
 )
 
-// hostKernel returns the running kernel version (equivalent to uname -r).
-func hostKernel() (string, error) {
+// HostKernel returns the running kernel version (equivalent to uname -r).
+func HostKernel() (string, error) {
 	var utsname syscall.Utsname
 	if err := syscall.Uname(&utsname); err != nil {
 		return "", fmt.Errorf("uname: %w", err)
 	}
 
+	// Utsname.Release is a fixed-size byte array; convert to string and trim
+	// the trailing NUL bytes.
 	buf := make([]byte, 0, len(utsname.Release))
 	for _, b := range utsname.Release {
 		if b == 0 {
